products/guardian_runtime/requirements.txt
# OMEGA Guardian Runtime — MVP
pydantic>=2.0.0
pytest>=7.0.0
PyYAML>=6.0
